fix: Update imports and dependencies for wagmiConfig. New API that includes Wagmi's createConfig and replaces getDefaultWallets and connectorsForWallets
